separated the cell lines per tissue pie chart into TissueCellsPieChart.js to be used on both Tissues and Cell Lines summary pages.
